refactor: UI 클래스 기능 분리
