--- src/main/java/racingcar/view/Messages.java
@@ -1,7 +1,0 @@
-package racingcar.view;
-
-public final class Messages {
-    public static final String INPUT_CARS_MESSAGE = "경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)";
-    public static final String INPUT_TRIAL_MESSAGE = "시도할 회수는 몇회인가요?";
-    public static final String RESULT_MESSAGE = "실행 결과";
-}
