+++ src/main/java/racingcar/view/car/CarNamesOutputView.java
@@ -1,0 +1,10 @@
+package racingcar.view.car;
+
+public class CarNamesOutputView {
+
+    private static final String INPUT_CARS_MESSAGE = "경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)";
+
+    public static void inputCarNames() {
+        System.out.println(INPUT_CARS_MESSAGE);
+    }
+}
