+++ src/main/java/racingcar/view/trial/TrialInputView.java
@@ -1,0 +1,9 @@
+package racingcar.view.trial;
+
+import camp.nextstep.edu.missionutils.Console;
+
+public class TrialInputView {
+    public static String getRacingTrial() {
+        return Console.readLine();
+    }
+}
