+++ src/main/java/racingcar/view/trial/TrialOutputView.java
@@ -1,0 +1,11 @@
+package racingcar.view.trial;
+
+public class TrialOutputView {
+
+
+    private static final String INPUT_TRIAL_MESSAGE = "시도할 회수는 몇회인가요?";
+
+    public static void inputTrial() {
+        System.out.println(INPUT_TRIAL_MESSAGE);
+    }
+}
