+++ src/main/java/racingcar/view/winner/RacingWinnerView.java
@@ -1,0 +1,11 @@
+package racingcar.view.winner;
+
+import racingcar.domain.winner.Winners;
+
+public class RacingWinnerView {
+
+    public static void showWinners(Winners winners) {
+        System.out.println();
+        System.out.println(winners.getWinnersMessage());
+    }
+}
